Move reduce() into helpers.js
