+++ fx.js
@@ -3,8 +3,7 @@ const fs = require('fs')
 const tty = require('tty')
 const blessed = require('@medv/blessed')
 const stringWidth = require('string-width')
-const { walk, log } = require('./helpers')
-const reduce = require('./reduce')
+const { walk, reduce, log } = require('./helpers')
 const print = require('./print')
 const search = require('./search')
 const config = require('./config')
